Extract quantization post operation into another file
